refactor: move form helpers into shared file
